+++ packages/babel-preset-expo/src/index.ts
@@ -89,6 +89,11 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
       require.resolve('@babel/plugin-proposal-object-rest-spread'),
       { loose: false },
     ]);
+  } else {
+    // This is added back on hermes to ensure the react-jsx-dev plugin (`@babel/preset-react`) works as expected when
+    // JSX is used in a function body. This is technically not required in production, but we
+    // should retain the same behavior since it's hard to debug the differences.
+    extraPlugins.push(require('@babel/plugin-transform-parameters'));
   }
 
   if (platformOptions.useTransformReactJSXExperimental != null) {